+++ changelog.md
@@ -24,6 +24,7 @@ Codion Change Log
 ### is.codion.swing.common.model
 - FilterComboBoxModel.ComboBoxSelection.value() and nullSelected() removed.
 - DefaultFilterTableSelection bug fixed, back to inheriting from DefaultListSelectionModel since some JTable keyboard actions rely on it.
+- DefaultFilterComboBoxModel.DefaultComboBoxItems.removeItem() now sets the selection to null if the selected item is removed, addItems() and removeItems() implemented.
 ### is.codion.swing.common.ui
 - TableConditionPanel bug fixed, clear control did not use TableConditionModel.clear().
 - SearchHighlighter refactored.